requirements.txt
Django~=4.0.4
djangorestframework~=3.13.1
numpy~=1.22.4
pytest~=7.1.2
pytz~=2022.1
pathlib~=1.0.1
Pygments~=2.12.0
coreschema~=0.0.4
requests~=2.27.1